+++ bindings.py
@@ -57,211 +57,520 @@ class User(Base):
 
 # Qt interface
 # # Mappers
-class QtDataMapper(object):
+class PQSFieldBinder(object):
+    """PyQt-SQLAclhemy (PQS) field binder
+
+    Binds a PyQt field object in a GUI (like a QLineEdit) to a SQLAlchemy
+    field in a model.
+
+    This allows to automatically update the SQLAlchemy field in the model when
+    the PyQt field object gets update. The SQLAlchemy field can have validators
+    so the GUI responds to them graphically.
+
+    This is just a base class to define a binder. Manual bindings have to be
+    set for each field in each UI created, overriding the required methods to
+    provide the autoupdate feature.
+
+    Attributes:
+        __metaclass__ (class): defines this as an abstract class
+        _gui_field (QWidget): PyQt object to get and set values from / to
+        _ui (PQSDataUi): UI that controls the editing features and holds the
+                         SQLAlchemy model
+        _name (str): Name of the field we are binding in the SQLAlchemy model
+        _autoupdate (bool): Controls if the SQLAlchemy will be updated
+                            automatically upon detecting a change in the PyQt
+                            field
+    """
+    CLASS_SUFFIX = "Binder"
+    """
+        str: suffix all derived class will have in their names so the name of
+             the binded field can be guessed in `__init__` method
+    """
+    COLOR_INVALID = "#f6989d"
+    """
+        str: background color to set to a PyQt field's if the displayed value
+             is invalid
+    """
+    COLOR_NOTVALIDATED = "#fff79a"
+    """
+        str: background color to set to a PyQt field's if the displayed value
+             has not been checked against validity yet
+    """
+    COLOR_VALID = "#c4df9b"
+    """
+        str: background color to set to a PyQt field's if the displayed value
+             is valid
+    """
+    VALIDITY_MARK_SYTLESHEET = "{0} {{ background-color: {1} }}"
+    """
+        str: stylesheet to set to a PyQt field to show graphically its validity
+             default changes background color
+             must format with:
+                1) the PyQt field class name:
+                    `self._gui_field.__class__.__name__`
+                2) the color the background will be set to
+    """
     __metaclass__ = ABCMeta
 
-    def __init__(self, field, ui, autoupdate=True):
-        self._field = field
-        self._ui = ui
-        self._name = self.__class__.__name__.lower().replace("mapper", "")
-        self._autoupdate = autoupdate
+    def __init__(self, gui_field, ui, name=None, autoupdate=True,
+                 connect_autoupdate=True):
+        """Initializes a PQSFieldBinder
 
+        Notes:
+            If name is not provided, will be guessed from the class name,
+            previously removing the suffix `CLASS_SUFFIX` and lowercasing it
+
+        Args:
+            gui_field (QWidget): PyQt object the SQLAlchemy model field will be
+                                 binded to
+            ui (class): UI that holds the SQAlchemy model and editing features
+            name (str): name of the field of the SQLAlchemy model we are
+                        binding
+            autoupdate (bool): whether to enable autoupdate (after connecting)
+            autoupdate_connect (bool): whether to connect the signal and slot
+                                       to enable autoconnect
+        """
+        self._gui_field = gui_field
+        self._ui = ui
+        # Autoguess name based on CLASS_SUFFIX
+        self._name = \
+            self.__class__.__name__.replace(self.CLASS_SUFFIX, "").lower() \
+            if name is None else name
+        self._autoupdate = autoupdate
+        # Autoupdate connect if needed
+        if connect_autoupdate:
+            self.connect_autoupdate()
+
+    # Control the binder features
     @property
     def autoupdate(self):
+        """Returns if the `autoupdate` feature is enabled"""
         return self._autoupdate
 
     def enable_autoupdate(self):
+        """Enables the `autoupdate` feature"""
         self._autoupdate = True
 
     def disable_autoupdate(self):
+        """Disables the `autoupdate` feature"""
         self._autoupdate = False
 
+    def connect_autoupdate(self):
+        """Connects the `gui_value_modified` signal to its trigger"""
+        self.signal_autoupdate.connect(self._on_autoupdate)
+
+    def disconnect_autoupdate(self):
+        """Disconnects the `gui_value_modified` signal from its trigger"""
+        self.signal_autoupdate.disconnect(self._on_autoupdate)
+
+    # Retrieve binder properties
     @property
     def name(self):
+        """Returns the name of the binded field in the SQLAlchemy model"""
         return self._name
 
     @property
     @abstractmethod
-    def qt_value(self):
-        pass
+    def signal_autoupdate(self):
+        """Returns the PyQt signal to detect a change if GUI value was modified
 
-    @qt_value.setter
-    @abstractmethod
-    def qt_value(self, new_value):
-        pass
+        Will be used to connect autoupdate signal with its event handler (slot)
 
-    @property
-    @abstractmethod
-    def signal(self):
+        It must be overwritten so the signal is defined"""
         pass
-
-    @abstractmethod
-    def disable(self):
-        pass
-
-    @abstractmethod
-    def enable(self):
-        pass
-
-    @property
-    def reset(self):
-        return None
 
     @property
     def model(self):
+        """Returns the SQLAlchemy model object"""
         return self._ui.model
 
     @property
+    def reset_value(self):
+        """Returns the reset value in a SQLAlchemy model field
+
+        The reset value stands for the value a field has before being
+        initialized. Therefore, when we want to remove a field's value, we'll
+        set it to this value.
+
+        In SQLAlchemy, this value is `None`. You can customize it if your model
+        requires specials needs but this is not recommended as your model has
+        to deal with `None` value as it's the SQLAlchemy default for non-set
+        values in a field
+        """
+        return None
+
+    # GUI properties: set and retrieve from GUI
+    @property
+    @abstractmethod
+    def gui_value(self):
+        """Returns the value of the PyQt field object
+
+        Must return something that can be set to the model field, providing
+        the needed conversions from Qt types to Python built-in types.
+
+        It will use one of the PyQt field's methods to retrieve a value, like
+        `text()` on a `QLineEdit` object.
+        """
+        pass
+
+    @gui_value.setter
+    @abstractmethod
+    def gui_value(self, new_value):
+        """Sets the value of a PyQt field object
+
+        Will be passed some value that may be in the SQLAlchemy model field
+        and it must be able to set it in the PyQt field.
+
+        It will use one of the PyQt field's methods to set a value, like
+        `setText()` on a `QLineEdit` object.
+
+        Notes:
+            The model value can be `None` if the model's field value has not
+            been set yet. The setter must set this properly and don't display
+            "None" in a `QLineEdit`, but "" instead. Take care of this when
+            writing fast conversions like `str(new_value)`
+        """
+        pass
+
+    def update_from_gui(self):
+        """Gets the GUI field value and sets it to the model
+
+        Uses `isValid` method. Tries to reset the value to the initial value
+        if can't be validated to avoid saving to the database invalid values.
+        """
+        value = self.gui_value
+        if not self.update(value):
+            self.reset()
+
+    def update_to_gui(self):
+        """Sets the GUI field value from the model's value
+
+        Marks also the field as not validated until the value is set
+        """
+        value = self.model_value
+        self.gui_value = value
+        self.mark_as_notvalidated(value)
+
+    # Model properties: set and retrieve from model
+    @property
     def model_value(self):
+        """Returns the value of the SQLAlchemy model's binded field"""
         return getattr(self.model, self.name)
 
     @model_value.setter
     def model_value(self, value):
-        raised_exception = None
+        """Sets the value of the SQLAlchemy model's binded field
+
+        Alias of `update`, without returning anything
+        """
+        self.update(value)
+
+    def update(self, value=None):
+        """Updates the binded model's field (and therefore also validates it)
+
+        Like `model_value(value)`, but controls exceptions, triggers events and
+        by default retrieves the value using the `gui_value` property
+
+        If no exception is triggered, the value therefore has been set to the
+        model and therefore has been validated if a validator existed. Then,
+        returns True. If an exception is triggered, returns False.
+
+        Notes:
+            Event-driven methods `self._on_invalid` and `self._on_valid` will
+            be triggered depending whether the operation succeeds or not.
+
+            Event-driven methods `self.on_valid` and `self.on_invalid` will be
+            triggered afterwards the previous ones (one or other depending the
+            validation result)
+
+        Args:
+            value (object): value to set to the model
+
+        Returns:
+            bool: True if PyQt's field value can be set to the SQLAlchemy model
+            field (therefore validates), False if not.
+        """
+        # Get value
+        value = self.gui_value if value is None else value
+        # Try to set it
         try:
             setattr(self.model, self.name, value)
-        except Exception as e:
-            raised_exception = e
-        self._ui.update_status()
-        if raised_exception is not None:
-            raise raised_exception
-
-    def isValid(self):
-        try:
-            # we must check what we see in order to prevent
-            # can't reset issues
-            # (we write 1, delete it, fails but 1 still there)
-            self.model_value = self.qt_value
-        except Exception as e:
-            print(" attribute %s value \"%s\" is invalid" % (
-                self.name, self.qt_value))
-            import traceback
-            traceback.print_exc()
-            self._on_invalid()
+        except Exception as exc:
+            self._on_invalid(value, exc)
             return False
         else:
-            print(" attribute %s is valid" % self.name)
-            self._on_valid()
+            self._on_valid(value)
             return True
+        finally:
+            self._ui.update_status()
 
-    def valid(self):
-        self._field.setStyleSheet('QLineEdit { background-color: %s }' %
-                                  STATES_COLORS[2])
+    def reset(self):
+        """Resets the SQLAlchemy model field's value
 
-    def schrodinger(self):
-        self._field.setStyleSheet('QLineEdit { background-color: %s }' %
-                                  STATES_COLORS[1])
+        Sets the field's value of the model to `reset_value` property value.
+        It may fail if the field can't be blank.
 
-    def invalid(self):
-        self._field.setStyleSheet('QLineEdit { background-color: %s }' %
-                                  STATES_COLORS[0])
+        Notes:
+            Event-driven methods `self._on_reset_success` and
+            `self._on_reset_fail` will be triggered depending whether the
+            operation succeeds or not. `self._on_reset` will always be called
+            at the end.
 
-    def _on_valid(self):
-        self.valid()
+            Event-driven methods `self.on_reset_success` and
+            `self.on_reset_fail` will be triggered depending whether the
+            operation succeeds or not. `self.on_reset` will always be called
+            at the end
 
-    def _on_invalid(self):
-        self.invalid()
-
-    def _on_update_fail(self, exc, value):
-        self._on_invalid()
-        print(" not valid: %s" % str(exc))
-
-    def _on_update_success(self, value):
-        self._on_valid()
-        print(" valid")
-
-    def _on_update(self, value):
-        print(" completed edit: model.%s = %s" % (self.name,
-              getattr(self._ui.model, self.name)))
-
-    def _on_update_reset_fail(self, exc, value):
-        print(" can't reset attribute: %s" % str(exc))
-
-    def _on_update_reset_success(self, value):
-        print(" reset attribute to %s" % self.reset)
-
-    def _on_update_reset(self, value):
-        print(" completed reset")
-
-    def updater(self):
-        if not self._autoupdate:
-            print("Autoupdate disabled")
-            return
-        print("Updating field \"%s\"" % self.name)
-        value = self.qt_value
-        failed = False
-        # try to set value
+        Returns:
+            bool: True if succeeds, False if not
+        """
         try:
-            self.model_value = value
+            self.model_value = self.reset_value
         except Exception as exc:
-            self._on_update_fail(exc, value)
-            failed = True
+            self._on_reset_fail(exc)
         else:
-            self._on_update_success(value)
-        # try to reset if failed
-        if failed:
-            try:
-                self.model_value = self.reset
-            except Exception as exc:
-                self._on_update_reset_fail(exc, value)
-            else:
-                self._on_update_reset_success(value)
-            finally:
-                self._on_update_reset(value)
-        # finished update
-        self._on_update(value)
+            self._on_reset_success()
+        finally:
+            self._on_reset()
+
+    # Control GUI: set GUI properties
+    @abstractmethod
+    def disable_edit(self):
+        """Disables the PyQt field so the user can't write in it
+
+        For instance, in a `QLineEdit`, the method would call
+        `setReadOnly(False)` on it
+        """
+        pass
+
+    def disable(self):
+        """Disables the PyQt field so the user can't write in it
+
+        Autoupdate will be also disabled
+        """
+        self.disable_edit()
+        self.disable_autoupdate()
+
+    @abstractmethod
+    def enable_edit(self):
+        """Enables the PyQt field so the user can write in it
+
+        For instance, in a `QLineEdit`, the method would call
+        `setReadOnly(True)` on it
+        """
+        pass
+
+    def enable(self):
+        """Enables the PyQt field so the user can write in it
+
+        Autoupdate will be also enabled
+        """
+        self.enable_edit()
+        self.enable_autoupdate()
+
+    def mark_as_valid(self, value):
+        """Marks the PyQt field as valid to the user, modifying it
+
+        Args:
+            value (object): value set and valid
+        """
+        self._gui_field.setStyleSheet(self.VALIDITY_MARK_SYTLESHEET.format(
+            self._gui_field.__class__.__name__,
+            self.COLOR_VALID)
+        )
+
+    def mark_as_notvalidated(self, value):
+        """Marks the PyQt field as intermediate to the user, modifying it
+
+        Intermediate status is for a field that has not been validated
+
+        Args:
+            value (object): value set and not checked against validity
+        """
+        self._gui_field.setStyleSheet(self.VALIDITY_MARK_SYTLESHEET.format(
+            self._gui_field.__class__.__name__,
+            self.COLOR_NOTVALIDATED)
+        )
+
+    def mark_as_invalid(self, value, exc):
+        """Marks the PyQt field as invalid to the user, modifying it
+
+        Args:
+            value (object); value checked as invalid
+            exc (Exception): exception triggered so it is marked as invalid
+        """
+        print("invalid")
+        self._gui_field.setStyleSheet(self.VALIDITY_MARK_SYTLESHEET.format(
+            self._gui_field.__class__.__name__,
+            self.COLOR_INVALID)
+        )
+
+    # Event handlers
+    # # Reset
+    # # # Fixed
+    def _on_reset_success(self):
+        """Event triggered when model's value has been reseted
+
+        Default is to call the user-defined event
+        """
+        self.on_reset_success()
+
+    def _on_reset_fail(self, exc):
+        """Event triggered when model's value has to be reseted but fails
+
+        Default is to call the user-defined event
+
+        Args:
+            exc (Exception): exception triggered so it couldn't be reset
+        """
+        self.on_reset_fail()
+
+    def _on_reset(self):
+        """Event triggered when model's value has tried to be reseted
+
+        Default is to call the user-defined event
+        """
+        self.on_reset()
+
+    # # # User-defined
+    def on_reset_success(self):
+        """User event triggered when model's value has been reseted
+
+        It triggers just after `_on_reset_success` has been called
+        """
+        pass
+
+    def on_reset_fail(self, exc):
+        """User event triggered when model's value has to be reseted but fails
+
+        It triggers just after `_on_reset_fail` has been called
+
+        Args:
+            exc (Exception): exception triggered so it couldn't be reset
+        """
+        pass
+
+    def on_reset(self):
+        """User event triggered when model's value has tried to be reseted
+
+        It triggers just after `_on_reset` has been called
+        """
+        pass
+
+    # # Validation
+    # # # Fixed
+    def _on_valid(self, value):
+        """Event triggered when the field value has been validated
+
+        Default is to mark the PyQt field as valid and call the user-defined
+        event
+
+        Args:
+            value (object): value set and valid
+        """
+        self.mark_as_valid(value)
+        self.on_valid(value)
+
+    def _on_invalid(self, value, exc):
+        """Event triggered when the field value has been invalidated
+
+        Default is to mark the PyQt field as invalid and call the user-defined
+        event
+
+        Args:
+            value (object); value checked as invalid
+            exc (Exception): exception triggered so it is marked as invalid
+        """
+        import traceback
+        traceback.print_exc()
+        self.mark_as_invalid(value, exc)
+        self.on_invalid(value, exc)
+
+    # # # User-defined
+    def on_valid(self, value):
+        """User event triggered when the field value has been validated
+
+        It triggers just after `_on_valid` has been called
+
+        Args:
+            value (object): value set and valid
+        """
+        pass
+
+    def on_invalid(self, value, exc):
+        """User event triggered when the field value has been invalidated
+
+        It triggers just after `_on_invalid` has been called
+
+        Args:
+            value (object); value checked as invalid
+            exc (Exception): exception triggered so it is marked as invalid
+        """
+        pass
+
+    # # Autoupdate-related
+    # # # Fixed events
+    def _on_autoupdate(self):
+        """Triggered when we must autoupdate
+
+        Calls `update_from_gui` if autoupdate is enabled
+
+        Slot to connect to defined `signal_autoupdate`
+        """
+        if self.autoupdate:
+            self.update_from_gui()
 
 
-class EmailMapper(QtDataMapper):
+class EmailBinder(PQSFieldBinder):
     @property
-    def qt_value(self):
-        get_value = self._field.text()
+    def gui_value(self):
+        get_value = self._gui_field.text()
         return get_value if len(get_value) else None
 
-    @qt_value.setter
-    def qt_value(self, new_value):
+    @gui_value.setter
+    def gui_value(self, new_value):
         set_value = new_value
         if new_value is None:
             set_value = ""
         elif not isinstance(new_value, str):
             set_value = str(set_value)
-        self._field.setText(set_value)
+        self._gui_field.setText(set_value)
 
     @property
-    def signal(self):
-        return self._field.textChanged
+    def signal_autoupdate(self):
+        return self._gui_field.textChanged
 
-    def disable(self):
-        self._field.setReadOnly(True)
+    def disable_edit(self):
+        self._gui_field.setReadOnly(True)
 
-    def enable(self):
-        self._field.setReadOnly(False)
+    def enable_edit(self):
+        self._gui_field.setReadOnly(False)
 
 
-class NumberMapper(QtDataMapper):
+class NumberBinder(PQSFieldBinder):
     @property
-    def qt_value(self):
-        get_value = self._field.text()
+    def gui_value(self):
+        get_value = self._gui_field.text()
         return get_value if len(get_value) else None
 
-    @qt_value.setter
-    def qt_value(self, new_value):
+    @gui_value.setter
+    def gui_value(self, new_value):
         set_value = new_value
         if new_value is None:
             set_value = ""
         elif not isinstance(new_value, str):
             set_value = str(set_value)
-        self._field.setText(set_value)
+        self._gui_field.setText(set_value)
 
     @property
-    def signal(self):
-        return self._field.textChanged
+    def signal_autoupdate(self):
+        return self._gui_field.textChanged
 
-    def disable(self):
-        self._field.setReadOnly(True)
+    def disable_edit(self):
+        self._gui_field.setReadOnly(True)
 
-    def enable(self):
-        self._field.setReadOnly(False)
+    def enable_edit(self):
+        self._gui_field.setReadOnly(False)
 
 
 # # Create widget
@@ -297,18 +606,20 @@ class ExampleDataUI():
         self.verticalLayout.addWidget(self.validateButton)
         self.verticalLayout.addWidget(self.statusLabel)
         # Define mappings
-        self.mappers = [EmailMapper(self.emailLineEdit, self),
-                        NumberMapper(self.numberLineEdit, self)]
-        # Create updaters and events
+        self.mappers = [EmailBinder(self.emailLineEdit, self),
+                        NumberBinder(self.numberLineEdit, self)]
+        # Fill GUI
         for mapper in self.mappers:
-            mapper.qt_value = mapper.model_value
-            mapper.signal.connect(functools.partial(mapper.updater))
-            mapper.schrodinger()
+            mapper.disable()
+            mapper.update_to_gui()
+            mapper.enable()
             status = sqla_inspect(self.model)
+            print(status.modified)
             if status.persistent and not status.modified:
+                # Do not update if persistent and clean
                 self.update_status()
             else:
-                mapper.isValid()
+                mapper.update()
         # Save
         self.saveButton.clicked.connect(self.save)
         self.refreshButton.clicked.connect(self.refresh)
@@ -318,28 +629,25 @@ class ExampleDataUI():
 
     def validate(self):
         for mapper in self.mappers:
-            mapper.disable()
-            mapper.isValid()
-            mapper.enable()
+            mapper.disable_edit()
+            mapper.update()
+            mapper.enable_edit()
 
     def refresh(self):
         SESSION.refresh(self.model)
-        self.update_from_model()
+        self.update_to_gui()
         self.update_status()
 
-    def update_from_model(self):
+    def update_to_gui(self):
         for mapper in self.mappers:
             mapper.disable()
-            mapper.disable_autoupdate()
-            mapper.qt_value = mapper.model_value
-            mapper.schrodinger()
-            mapper.enable_autoupdate()
+            mapper.update_to_gui()
             mapper.enable()
 
     def save(self):
         print("Try to save")
         for mapper in self.mappers:
-            if not mapper.isValid():
+            if not mapper.update():
                 QMessageBox.warning(self.parent, "Form invalid",
                                     "Check and try again")
                 return